requirements.txt
beautifulsoup4==4.10.0
black==22.3.0
dateparser==1.1.1
isort==5.8.0
peewee==3.14.4
Pillow==9.0.1
pre-commit-hooks==4.0.1
pre-commit==2.13.0
py-trello==0.18.0
python-telegram-bot==13.6
requests==2.26.0
